tom: make setup-server rerunnable

+++ setup-server.sh
@@ -1,30 +1,33 @@
 #!/usr/bin/env bash
 environment=$1
 
+public_files_base="https://bitbucket.org/tomburnell/table1_public_files/raw/b21d88ff949b9824328f84b9a4aade0d3a950dfc"
 
-echo "PATH=/command:\$PATH" >> $HOME/.profile
-echo "export NODE_ENV=integration" >> /etc/profile
-echo "export PATH=/opt/local/bin:$PATH" >> /etc/profile
+if [ -z "$environment" ]; then
+    echo "USAGE: $0 <environment>"
+    exit 2
+fi
 
+if ! cat $HOME/.profile | grep  '/command' > /dev/null; then echo "PATH=/command:\$PATH" >> $HOME/.profile;  fi
+if [ cat /etc/profile | grep 'NODE_ENV' ];then echo "export NODE_ENV=integration" >> /etc/profile; fi
+if [ cat /etc/profile | grep '/opt/local/bin' ];then echo "export PATH=/opt/local/bin:$PATH" >> /etc/profile; fi
 
 mkdir -p /service
 if [ ! -d /command ]; then
     cd /var/tmp
-    curl -O  https://bitbucket.org/tomburnell/table1_public_files/raw/b21d88ff949b9824328f84b9a4aade0d3a950dfc/dt.tar.gz
-
+    curl -O  "$public_files_base/dt.tar.gz"
     cd /
     tar -xzvf /var/tmp/dt.tar.gz
     echo "SV:123456:respawn:/command/svscanboot" >> /etc/inittab
     ps -ef | grep init | grep -v grep | awk '{print $2}' | xargs kill -HUP
 fi
 
-
 svcadm disable /mailonline/MOLstatsd
 cd /var/tmp
 if [ -f statsd.tar.gz ]; then 
-  rm statsd.tar.gz
+    rm statsd.tar.gz
 fi
-curl -O https://bitbucket.org/tomburnell/table1_public_files/raw/b21d88ff949b9824328f84b9a4aade0d3a950dfc/statsd.tar.gz
+curl -O "$public_files_base/statsd.tar.gz"
 rm -rf /opt/molswf/statsd
 mkdir -p /opt/molsfw/statsd
 cd /opt/molsfw/statsd
